move setupLocale to class Oven
